+++ NEWS.md
@@ -1,3 +1,5 @@
+# ameras (development version)
+
 # ameras 0.2.0
 
 ## Breaking changes