+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * Shortened column names `CI.lowerbound` and `CI.upperbound` in the `summary()` output to `CI.lower` and `CI.upper`, respectively.
 
+* Substantially increased ERC computation speed for the `clogit` family.
+
+
 # ameras 0.2.0
 
 ## Breaking changes